requirements-dev.txt
-r requirements.txt

django-debug-toolbar==3.5.0